split into separate test
